Test restructuring
* Restructured smoke tests and fuzz tests to reduce repetition
* Increased test coverage to include proper testing for hybrid handshakes
