+++ fuzz/fuzz_targets/hybrid_handshake_read.rs
@@ -1,0 +1,41 @@
+#![no_main]
+
+use clatter::constants::MAX_MESSAGE_LEN;
+use clatter::crypto::cipher::{AesGcm, ChaChaPoly};
+use clatter::crypto::dh::X25519;
+use clatter::crypto::hash::{Blake2b, Blake2s, Sha256, Sha512};
+use clatter::crypto::kem::pqclean_ml_kem::MlKem1024;
+use clatter::crypto::kem::rust_crypto_ml_kem::MlKem512;
+use clatter::traits::{Cipher, Dh, Hash, Handshaker, Kem};
+use clatter_fuzz::{hybrid_handshake_patterns, setup_hybrid_handshake};
+use libfuzzer_sys::fuzz_target;
+
+fuzz_target!(|data: &[u8]| {
+    verify_with::<X25519, MlKem1024, MlKem512, AesGcm, Sha256>(data);
+    verify_with::<X25519, MlKem1024, MlKem512, AesGcm, Sha512>(data);
+    verify_with::<X25519, MlKem1024, MlKem512, AesGcm, Blake2b>(data);
+    verify_with::<X25519, MlKem1024, MlKem512, AesGcm, Blake2s>(data);
+    verify_with::<X25519, MlKem1024, MlKem512, ChaChaPoly, Sha256>(data);
+    verify_with::<X25519, MlKem1024, MlKem512, ChaChaPoly, Sha512>(data);
+    verify_with::<X25519, MlKem1024, MlKem512, ChaChaPoly, Blake2b>(data);
+    verify_with::<X25519, MlKem1024, MlKem512, ChaChaPoly, Blake2s>(data);
+});
+
+fn verify_with<DH: Dh, EKEM: Kem, SKEM: Kem, C: Cipher, H: Hash>(data: &[u8]) {
+    let handshakes = hybrid_handshake_patterns();
+
+    for pattern in handshakes {
+        let mut alice_buf = [0u8; MAX_MESSAGE_LEN];
+        let mut bob_buf = [0u8; MAX_MESSAGE_LEN];
+
+        let (mut alice, mut bob) = setup_hybrid_handshake::<DH, EKEM, SKEM, C, H>(&pattern);
+
+        // Verify Alice only if the pattern is not one-way
+        if !pattern.is_one_way() {
+            let _ = alice.write_message(&[], &mut alice_buf).unwrap();
+            let _ = alice.read_message(data, &mut alice_buf);
+        }
+        let _ = bob.read_message(data, &mut bob_buf);
+    }
+}
+
