+++ fuzz/fuzz_targets/hybrid_transport.rs
@@ -2,33 +2,33 @@
 
 use clatter::constants::MAX_MESSAGE_LEN;
 use clatter::crypto::cipher::{AesGcm, ChaChaPoly};
+use clatter::crypto::dh::X25519;
 use clatter::crypto::hash::{Blake2b, Blake2s, Sha256, Sha512};
 use clatter::crypto::kem::pqclean_ml_kem::MlKem1024;
 use clatter::crypto::kem::rust_crypto_ml_kem::MlKem512;
-use clatter::traits::{Cipher, Hash, Kem};
-use clatter_fuzz::{complete_handshake, pq_handshake_patterns, setup_pq_handshake};
+use clatter::traits::{Cipher, Dh, Hash, Kem};
+use clatter_fuzz::{complete_handshake, hybrid_handshake_patterns, setup_hybrid_handshake};
 use libfuzzer_sys::fuzz_target;
 
 fuzz_target!(|data: &[u8]| {
-    // TODO: generate all combinations
-    verify_with::<MlKem1024, MlKem512, AesGcm, Sha256>(data);
-    verify_with::<MlKem1024, MlKem512, AesGcm, Sha512>(data);
-    verify_with::<MlKem1024, MlKem512, AesGcm, Blake2b>(data);
-    verify_with::<MlKem1024, MlKem512, AesGcm, Blake2s>(data);
-    verify_with::<MlKem1024, MlKem512, ChaChaPoly, Sha256>(data);
-    verify_with::<MlKem1024, MlKem512, ChaChaPoly, Sha512>(data);
-    verify_with::<MlKem1024, MlKem512, ChaChaPoly, Blake2b>(data);
-    verify_with::<MlKem1024, MlKem512, ChaChaPoly, Blake2b>(data);
+    verify_with::<X25519, MlKem1024, MlKem512, AesGcm, Sha256>(data);
+    verify_with::<X25519, MlKem1024, MlKem512, AesGcm, Sha512>(data);
+    verify_with::<X25519, MlKem1024, MlKem512, AesGcm, Blake2b>(data);
+    verify_with::<X25519, MlKem1024, MlKem512, AesGcm, Blake2s>(data);
+    verify_with::<X25519, MlKem1024, MlKem512, ChaChaPoly, Sha256>(data);
+    verify_with::<X25519, MlKem1024, MlKem512, ChaChaPoly, Sha512>(data);
+    verify_with::<X25519, MlKem1024, MlKem512, ChaChaPoly, Blake2b>(data);
+    verify_with::<X25519, MlKem1024, MlKem512, ChaChaPoly, Blake2s>(data);
 });
 
-fn verify_with<EKEM: Kem, SKEM: Kem, C: Cipher, H: Hash>(data: &[u8]) {
-    let handshakes = pq_handshake_patterns();
+fn verify_with<DH: Dh, EKEM: Kem, SKEM: Kem, C: Cipher, H: Hash>(data: &[u8]) {
+    let handshakes = hybrid_handshake_patterns();
 
     for pattern in handshakes {
         let mut alice_buf = [0u8; MAX_MESSAGE_LEN];
         let mut bob_buf = [0u8; MAX_MESSAGE_LEN];
 
-        let (alice, bob) = setup_pq_handshake::<EKEM, SKEM, C, H>(&pattern);
+        let (alice, bob) = setup_hybrid_handshake::<DH, EKEM, SKEM, C, H>(&pattern);
 
         // Complete handshake
         let (mut alice, mut bob) = complete_handshake(alice, bob);
@@ -48,3 +48,4 @@ fn verify_with<EKEM: Kem, SKEM: Kem, C: Cipher, H: Hash>(data: &[u8]) {
         }
     }
 }
+
